Factor out the JS compiler-specific logic from RhinoEvaluator into separate class
The RhinoCompiler class now contains all the JS compiler-specific methods to setup the Rhino scope/context environment in a way that makes sense for JS compiler use-cases. Basically, allows registering the compiler ahead-of-time with a name, as well as passing in the source code to be compiled with compiler options. Both the string of source code and the map of compiler options are stored in standard fields inside the __Screwdriver object available to the JS compiler program when run.