Centralize run-time configuration state, start work on includes/excludes
The LESS compiler, unlike SASS, does not have a built-in concept of "partials," so every file it gets it compiles. This does not work for files that are intended as partials (i.e., they know that other dependencies will be loaded in a parent script in which they themselves are loaded). Without the ability to white-list which LESS files get compiled, folks would have to duplicate @import statements in multiple files. I'd like not to add in the concept of partials like other build tools do, but rather give a simple mechanism to tell the LESS compiler which files to worry about as top-level compilation units.