Include JS dependencies for RequireJS optimizer, include working config
The config hardcoded-build.js will work if run from the root of Screwdriver (see instructions in main README). The requirejs-build.js file may be only useful in part, since we'll need to require that users specify the "main" modules that should be targeted for optimization, in which case we'll be generating these r.js config objects dynamically.