Add simplistic API layer in Clojure to clean up project configs
Limited to the fields used by a current project, this provides a function build-config that takes in a map of configuration options and does the right things with them. This limits the number of Java-level things that need to be done in end-user project configurations for Screwdriver and should eventually be fleshed out to cover all configuration options available in Screwdriver.